feedback review system

+++ feedback/views.py
@@ -410,7 +410,7 @@ def review(request):
                     for a in answers:
                         number +=1.0
                         add += a.rating
-                    average=str(add/number)            
+                    average=Decimal(str(add/number))            
                     saveavg = Answeravg(question=q,owner=curr_userprofile,avg=average, num=number)
                     saveavg.save()
         averages = Answeravg.objects.filter(owner=curr_userprofile)
